modify
git-svn-id: https://svn.baidu.com/myspace/sep/chenghuige/melt-train@184282 c4ccc3fd-b13d-49bf-a705-f618d8f4424a

+++ include/Trainers/FastRank/FastRankArguments.h
@@ -21,13 +21,13 @@ struct FastRankArguments
 	int numTrees = 100;
 	int numLeaves = 20;
 	int minInstancesInLeaf = 10;
-	double learningRate = 0.2;
+	Float learningRate = 0.2;
 
 	bool calibrateOutput = true; //calibrate| use calibrator to gen probability?
 	string calibratorName = "sigmoid"; //calibrator| sigmoid/platt naive pav @TODO 移动到Trainer父类处理
 
 	int maxBins = 255; //mb|Maximum number of distinct values (bins) per feature
-	double sparsifyRatio = 0.3;//sr|if not big data (large instances num, large feature num can set 0 so to be all dense) that will make running faster
+	Float sparsifyRatio = 0.3;//sr|if not big data (large instances num, large feature num can set 0 so to be all dense) that will make running faster
 
 	unsigned randSeed = 0x7b; //rs|controls wether the expermient can reproduce, 0 means not reproduce rngSeed
 	bool randomStart = false; //rst|Training starts from random ordering (determined by /r1)
@@ -39,25 +39,25 @@ struct FastRankArguments
 	bool affineRegressionTrees = false; //art|Use affine regression trees
 	bool allowDummyRootSplits = true; //dummies|When a root split is impossible, construct a dummy empty tree rather than fail
 
-	double baggingTrainFraction = 0.7;//bagfrac|Percentage of training queries used in each bag
+	Float baggingTrainFraction = 0.7;//bagfrac|Percentage of training queries used in each bag
 	bool bestStepRankingRegressionTrees = false; //bsr|Use best ranking step trees
 
-	double entropyCoefficient = 0; //e|The entropy (regularization) coefficient between 0 and 1
+	Float entropyCoefficient = 0; //e|The entropy (regularization) coefficient between 0 and 1
 
 	int derivativesSampleRate = 1; //dsr|same each query 1 in k times in the GetDerivatives function
 
-	double smoothing = 0.0; //s|Smoothing paramter for tree regularization
+	Float smoothing = 0.0; //s|Smoothing paramter for tree regularization
 
 	bool compressEnsemble = false; //cmp|Compress the tree Ensemble
 
-	double featureFirstUsePenalty = 0;
-	double featureReusePenalty = 0;
-	double softmaxTemperature = 0;
+	Float featureFirstUsePenalty = 0;
+	Float featureReusePenalty = 0;
+	Float softmaxTemperature = 0;
 
-	double featureFraction = 1;
-	double splitFraction = 1;
+	Float featureFraction = 1;
+	Float splitFraction = 1;
 	bool filterZeroLambdas = false;
-	double gainConfidenceLevel = 0;
+	Float gainConfidenceLevel = 0;
 };
 
 typedef shared_ptr<FastRankArguments> FastRankArgumentsPtr;
